refactor(state/workflows): split helpers.ts (1496 LOC) into focused sub-modules (ADR 0005 §4)
helpers.ts is now a barrel re-export; each new file has a clear single responsibility:
- workflow-node-display.ts   — value-section preview + override badge text
- workflow-graph-traversal.ts — findCycleNodes, reachableFrom*, connectedComponents
- workflow-graph-analysis.ts  — GraphHealth, inspectGraph, workflowDepth, findIsolated/Sources/Targets
- workflow-execution.ts       — flightFromNode, topologicalOrder
- workflow-layout.ts          — placeNewNode, autoLayout, compactPositions, nodeBounds, cloneNodeAt
- workflow-lookups.ts         — collection queries, name helpers, summariseHealth
- workflow-health.ts          — validateConnection, cleanupDanglingEdges, NodeIssue helpers
- workflow-tags.ts            — extractAllTags, tagCount, unusedTags, workflowsByTag
- workflow-search.ts          — searchNodes, searchWorkflows, formatRelativeTime
- workflow-serialisation.ts   — duplicateWorkflow, mergeWorkflows, parseImportedWorkflow

No behaviour changes. All 629 tests pass. Typecheck green.
